# docs/openapi-instant-meetings.yaml
openapi: 3.0.3
info:
  title: API pública — Salas e reuniões
  version: 3.2.1
  description: |
    API para plataformas terceirizadas criarem **salas-template** (personalização reutilizável)
    e **reuniões** (sessões com URL de entrada `/m/{slug}`).

    ## Fluxos

    1. **Sala padrão** — `POST /api/v1/rooms`  
       Cria um template de marca. O `name` identifica a sala; a personalização
       (identidade, paleta, avançado) é opcional e usa defaults da plataforma
       quando omitida.

    2. **Reunião a partir de sala** — `POST /api/v1/rooms/{room_id}/meetings`  
       Cria uma reunião com o título fornecido e snapshot da marca da sala.

    3. **Reunião instantânea** — `POST /api/v1/instant-meetings`  
       Cria uma reunião sem sala. Campos de identidade/paleta/avançado são
       opcionais (default = UI predefinida do owner ou tema da plataforma).

    4. **Integração LMS/CRM (privada + retorno)**  
       - Crie com `access_policy: invite` (ativa `wait_for_host` por omissão).  
       - Partilhe `url` com participantes (ficam na sala de espera até o anfitrião).  
       - Redirecione o anfitrião para `host_url` (concede privilégios de host sem login).  
       - Após leave/end, participantes regressam a `redirect_after_meet`.  
       - Informe `webhook_url` para receber resumo, transcrição, tarefas e gravação.

    ## Webhooks de saída (webhook_url)

    Campo opcional em salas e reuniões: URL **http(s) absoluta** para onde o
...

...
      são enviados para lá.

...

...

...

...

...

...

...
servers:
...
paths:
  /api/v1/rooms:
    post:
      operationId: createRoom
      summary: Criar sala padrão (template de personalização)
      description: |
        Cria uma sala-template em `/r/{slug}`. **Não cria reunião.**

        O campo `name` serve apenas para identificar o template. O título que
        aparece na reunião deve ser definido depois em
        `POST /api/v1/rooms/{room_id}/meetings`.

        Grupos `identity`, `palette` e `advanced` são opcionais. Campos omitidos
        usam o padrão da plataforma (`theme_preset: sky`, wordmark/lobby a partir
        de `name`, subtítulo "Powered by OpenMeet").
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
            examples:
              minimal:
                summary: Só o nome (tema padrão da plataforma)
                value:
                  name: Template Acme
                  external_id: "partner-user-123"
              withWebhook:
                summary: Nome + destino dos artefatos da reunião
                value:
                  name: Template Acme
                  external_id: "partner-user-123"
                  webhook_url: "https://lms.example.com/hooks/openmeet"
              fullBrand:
                summary: Identidade + paleta + avançado
                value:
                  name: Template Acme
                  external_id: "partner-user-123"
                  access_policy: public
                  identity:
                    logo_url: "https://cdn.example.com/logo.png"
                    wordmark: Acme
                    lobby_subtitle: Reuniões Acme
                  palette:
                    theme_preset: emerald
                    bg_animation: aurora
                    bg_animation_speed: 4
                  advanced:
                    font_family: "Inter, system-ui, sans-serif"
      responses:
        "201":
          description: Sala-template criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateRoomResponse"
              example:
                room_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                name: Template Acme
                slug: "x7k9m2pqab"
                url: "https://openmeet.chronos.com.pt/r/x7k9m2pqab"
                join_path: "/r/x7k9m2pqab"
                access_policy: public
                webhook_url: null
                brand:
                  identity:
                    lobby_title: Template Acme
                    lobby_subtitle: Powered by OpenMeet
                    logo_url: null
                    wordmark: Template Acme
                  palette:
                    theme_preset: sky
                    primary_color: "#0ea5e9"
                    secondary_color: "#38bdf8"
                    tertiary_color: "#818cf8"
                    background: "#0b1020"
                  advanced:
                    font_family: "Inter, system-ui, sans-serif"
                    favicon_url: null
                    custom_css: null
        "400":
          description: Corpo inválido ou falha ao resolver owner
        "401":
          description: Autenticação em falta ou inválida
  /api/v1/rooms/{room_id}/meetings:
    post:
      operationId: createMeetingFromRoom
      summary: Criar reunião a partir de sala predefinida
      description: |
        Cria uma reunião independente (`/m/{slug}`) usando a marca da sala como
        snapshot. O **título da reunião é obrigatório** e vem da plataforma
        terceirizada — não usa o `name` da sala.

        Sem override de personalização neste endpoint (a marca vem da sala).
      security:
        - bearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID da sala-template criada em POST /api/v1/rooms
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingFromRoomRequest"
            examples:
              basic:
                summary: Título obrigatório
                value:
                  title: Comercial 8 set
              withWebhook:
                summary: Título + webhook da reunião (override da sala)
                value:
                  title: Comercial 8 set
                  webhook_url: "https://crm.example.com/hooks/meetings/comercial"
              withTimeout:
                summary: Título + timeout
                value:
                  title: Kickoff parceiro
                  access_policy: public
                  empty_timeout_sec: 300
      responses:
        "201":
          description: Reunião criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMeetingResponse"
              example:
                meeting_id: "8c3e2a1b-5717-4562-b3fc-2c963f66afa6"
                slug: "m7k9m2pqab"
                url: "https://openmeet.chronos.com.pt/m/m7k9m2pqab"
                join_path: "/m/m7k9m2pqab"
                host_url: "https://openmeet.chronos.com.pt/api/v1/meetings/8c3e2a1b-5717-4562-b3fc-2c963f66afa6/enter?token=…"
                host_path: "/api/v1/meetings/8c3e2a1b-5717-4562-b3fc-2c963f66afa6/enter?token=…"
                access_policy: public
                title: Comercial 8 set
                brand_room_id: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                empty_timeout_sec: 300
                redirect_after_meet: null
                external_invite_url: null
                webhook_url: null
                wait_for_host: false
        "400":
          description: Corpo inválido (ex. title em falta)
        "401":
          description: Autenticação em falta ou inválida
        "404":
          description: Sala não encontrada
...
  /api/v1/meetings/{meeting_id}/enter:
    get:
      operationId: enterMeetingAsHost
      summary: Entrada pré-redirect do anfitrião
      description: |
        Endpoint de browser (sem Bearer). Valida o `token` emitido em `host_url`
        na criação da reunião, grava um cookie httpOnly de anfitrião e redireciona
        (302) para `/m/{slug}` com privilégios de host (encerrar reunião, etc.).

        Use este link para o utilizador que deve entrar como anfitrião a partir
        da plataforma terceirizada.
      parameters:
        - name: meeting_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: Token secreto devolvido apenas via host_url na criação
        - name: display_name
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
      responses:
        "302":
          description: Cookie de host definido; redirect para a lobby `/m/{slug}`
        "403":
          description: Token inválido
        "404":
          description: Reunião não encontrada
        "410":
          description: Reunião já encerrada
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token gerado em Admin → API (ou MEET_MCP_TOKEN)
  schemas:
    PaintStop:
      type: object
      required: [color, at]
      properties:
        color:
          type: string
          example: "#10b981"
        at:
          type: number
          minimum: 0
          maximum: 100
          description: Posição do stop (0–100)
    PaintGradient:
      type: object
      required: [type, angle, stops]
      properties:
        type:
          type: string
          enum: [linear, radial]
        angle:
          type: number
          minimum: 0
          maximum: 360
        stops:
          type: array
          minItems: 2
          maxItems: 5
          items:
            $ref: "#/components/schemas/PaintStop"
    PaintToken:
      type: object
      nullable: true
      required: [mode, solid]
      properties:
        mode:
          type: string
          enum: [solid, gradient]
        solid:
          type: string
          description: Cor sólida (também fallback quando mode=gradient)
          example: "#0ea5e9"
        gradient:
          $ref: "#/components/schemas/PaintGradient"
    BrandIdentity:
      type: object
      description: |
        Tab Identidade. Todos os campos são opcionais.
        Defaults: lobby_title/wordmark = nome da sala ou título da reunião;
        lobby_subtitle = "Powered by OpenMeet"; logo_url = null.
      properties:
        lobby_title:
          type: string
          description: Título mostrado no lobby
        lobby_subtitle:
          type: string
          description: Subtítulo do lobby (default "Powered by OpenMeet")
        logo_url:
          type: string
          nullable: true
          description: URL absoluta http(s) ou caminho relativo (/brand-assets/…)
        wordmark:
          type: string
          nullable: true
          maxLength: 120
          description: Texto exibido quando não há logo
    BrandPalette:
      type: object
      description: |
        Tab Paleta. Todos os campos são opcionais.
        Default theme_preset = sky (#0ea5e9 / #38bdf8 / #818cf8, fundo #0b1020).
        Se só theme_preset for enviado, as cores do preset são aplicadas.
      properties:
        theme_preset:
          type: string
          enum: [sky, violet, indigo, emerald, rose, amber]
          description: Preset de tema (default sky)
          default: sky
        primary_color:
          type: string
          example: "#0ea5e9"
        secondary_color:
          type: string
          example: "#38bdf8"
        tertiary_color:
          type: string
          example: "#818cf8"
        background:
          type: string
          example: "#0b1020"
          description: Fundo sólido (default #0b1020)
        primary_paint:
          $ref: "#/components/schemas/PaintToken"
        secondary_paint:
          $ref: "#/components/schemas/PaintToken"
        tertiary_paint:
          $ref: "#/components/schemas/PaintToken"
        background_paint:
          $ref: "#/components/schemas/PaintToken"
        pattern_url:
          type: string
          nullable: true
          description: URL do padrão de fundo
        pattern_size_mode:
          type: string
          nullable: true
          enum: [percent, fixed]
          default: percent
        pattern_size:
          type: integer
          nullable: true
          minimum: 1
          maximum: 512
          default: 24
        pattern_tint:
          type: string
          nullable: true
          enum: [none, primary, secondary, tertiary, custom]
          default: none
        pattern_tint_color:
          type: string
          nullable: true
        pattern_tint_opacity:
          type: integer
          nullable: true
          minimum: 0
          maximum: 100
          default: 55
        bg_animation:
          type: string
          nullable: true
          enum: [none, wave, beam, aurora, pulse]
          default: none
        bg_animation_speed:
          type: integer
          nullable: true
          minimum: 1
          maximum: 10
          default: 3
    BrandAdvanced:
      type: object
      description: |
        Tab Avançado. Todos os campos são opcionais.
        Default font_family = "Inter, system-ui, sans-serif".
      properties:
        font_family:
          type: string
          example: "Inter, system-ui, sans-serif"
          default: "Inter, system-ui, sans-serif"
        favicon_url:
          type: string
          nullable: true
        custom_css:
          type: string
          nullable: true
          description: CSS injetado na página da sala/reunião
    BrandPublic:
      type: object
      description: Marca aplicada (resposta), agrupada como nas tabs do editor
      properties:
        identity:
          $ref: "#/components/schemas/BrandIdentity"
        palette:
          $ref: "#/components/schemas/BrandPalette"
        advanced:
          $ref: "#/components/schemas/BrandAdvanced"
...
    CreateRoomRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Nome de identificação do template (não é o título da reunião)
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          minLength: 2
          maxLength: 64
          description: Slug opcional; gerado automaticamente se omitido
        access_policy:
          type: string
          enum: [public, members, invite]
          default: public
        board_id:
          type: string
        owner_identity_id:
          type: string
          format: uuid
          description: UUID interno do owner (utilizador OpenMeet)
        owner_user_id:
          type: string
          format: uuid
          description: Alias de owner_identity_id
        external_id:
          type: string
          description: |
            ID externo do utilizador na plataforma parceira. Se omitido
            (com token de Admin), usa o admin que gerou o token.
        identity:
          $ref: "#/components/schemas/BrandIdentity"
        palette:
          $ref: "#/components/schemas/BrandPalette"
        advanced:
          $ref: "#/components/schemas/BrandAdvanced"
        webhook_url:
          $ref: "#/components/schemas/WebhookUrl"
        captions_enabled:
          $ref: "#/components/schemas/CaptionsEnabled"
        transcription_enabled:
          $ref: "#/components/schemas/TranscriptionEnabled"
        summary_enabled:
          $ref: "#/components/schemas/SummaryEnabled"
    CreateRoomResponse:
      type: object
      required: [room_id, name, slug, url, join_path]
      properties:
        room_id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        url:
          type: string
          format: uri
          description: URL do template (/r/{slug})
        join_path:
          type: string
          example: /r/x7k9m2pqab
        access_policy:
          type: string
        webhook_url:
          $ref: "#/components/schemas/WebhookUrl"
        captions_enabled:
          type: boolean
          description: Valor gravado na sala (modelo das próximas reuniões)
        transcription_enabled:
          type: boolean
        summary_enabled:
          type: boolean
          description: Sempre false quando transcription_enabled=false
        brand:
          $ref: "#/components/schemas/BrandPublic"
    CreateMeetingFromRoomRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Título da reunião (obrigatório — definido pela plataforma)
        access_policy:
          type: string
          enum: [public, members, invite]
          description: Se omitido, herda da sala
        board_id:
          type: string
        empty_timeout_sec:
          type: integer
          minimum: 60
          maximum: 86400
          description: |
            Segundos após o último participante sair até a sala LiveKit terminar.
            Omitir para usar o default do servidor (LIVEKIT_EMPTY_TIMEOUT_SEC).
        redirect_after_meet:
          type: string
          format: uri
          nullable: true
          description: URL http(s) absoluta para onde enviar o utilizador após leave/end
        external_invite_url:
          $ref: "#/components/schemas/ExternalInviteUrl"
        webhook_url:
          $ref: "#/components/schemas/WebhookUrl"
        wait_for_host:
          type: boolean
          description: |
            Se true, participantes ficam na lobby até existir um anfitrião na sala.
            Por omissão true quando access_policy=invite.
        captions_enabled:
          $ref: "#/components/schemas/CaptionsEnabled"
        transcription_enabled:
          $ref: "#/components/schemas/TranscriptionEnabled"
        summary_enabled:
          $ref: "#/components/schemas/SummaryEnabled"
        live_stream_enabled:
          type: boolean
          description: |
            Permite que anfitriões transmitam esta reunião ao vivo (YouTube/RTMP).
            Omitido = LIVE_STREAM_MEETING_DEFAULT da instância ("on" por padrão).
            O interruptor da instância (/admin → Transmissão) continua valendo.
...
    CreateMeetingResponse:
      type: object
      required: [meeting_id, slug, url, join_path]
      properties:
        meeting_id:
          type: string
          format: uuid
        slug:
          type: string
        url:
          type: string
          format: uri
          description: URL de entrada dos participantes (/m/{slug})
        join_path:
          type: string
          example: /m/x7k9m2pqab
        host_url:
          type: string
          format: uri
          nullable: true
          description: |
            URL de entrada do anfitrião (GET …/enter?token=…). Concede host
            sem conta na plataforma de videoconferência.
        host_path:
          type: string
          nullable: true
          example: /api/v1/meetings/{id}/enter?token=…
        access_policy:
          type: string
        title:
          type: string
        brand_room_id:
          type: string
          format: uuid
          nullable: true
          description: ID da sala-template usada como marca (null se instantânea pura)
        empty_timeout_sec:
          type: integer
          description: Timeout efetivo de sala vazia em segundos
        redirect_after_meet:
          type: string
          format: uri
          nullable: true
        external_invite_url:
          $ref: "#/components/schemas/ExternalInviteUrl"
        webhook_url:
          $ref: "#/components/schemas/WebhookUrl"
        wait_for_host:
          type: boolean
          description: Participantes aguardam o anfitrião na lobby
        captions_enabled:
          type: boolean
          description: Valor efetivo da reunião (explícito, herdado da sala ou true)
        transcription_enabled:
          type: boolean
        summary_enabled:
          type: boolean
          description: |
            Sempre false quando transcription_enabled=false. Com false, o resumo
            fica com status "disabled" e os webhooks summary.ready e
            tasks.generated não são enviados.
        live_stream_enabled:
          type: boolean
          nullable: true
          description: |
            Valor gravado na reunião. null = segue LIVE_STREAM_MEETING_DEFAULT
            da instância.
    CaptionsEnabled:
      type: boolean
      description: |
        Legendas ao vivo na reunião. Default true. Em POST /api/v1/rooms/{room_id}/meetings
        e em instant-meetings com room_id, omitir herda o valor da sala.
        Com captions_enabled=false e transcription_enabled=false o agente de IA
        não entra na reunião (sem legenda, transcrição nem copiloto por voz).
    TranscriptionEnabled:
      type: boolean
      description: |
        Gravação da transcrição (transcrição completa, copiloto e webhook
        transcript.ready). Default true; omitir herda da sala quando houver.
        Com false, summary_enabled é forçado para false.
    SummaryEnabled:
      type: boolean
      description: |
        Resumo e tarefas gerados no fim da reunião. Default true; omitir herda da
        sala quando houver. Depende da transcrição: com transcription_enabled=false
        o valor é normalizado para false sem erro.
    WebhookUrl:
      type: string
      format: uri
      nullable: true
      maxLength: 2000
      description: |
        URL http(s) absoluta para onde enviar os artefatos da reunião
        (resumo, transcrição completa, tarefas e link da gravação).
        Não é redirect de browser — isso é `redirect_after_meet`.
    ExternalInviteUrl:
      type: string
      format: uri
      nullable: true
      maxLength: 2000
      description: |
        URL http(s) absoluta copiada pela ação de convite dentro da reunião.
        Se omitida ou null, o OpenMeet copia a URL atual da reunião.
...