# src/app/api/v1/rooms/[room_id]/meetings/route.ts
import { NextRequest, NextResponse } from "next/server";
import { eq } from "drizzle-orm";
import { z } from "zod";
import { db } from "@/db";
import { rooms } from "@/db/schema";
import { parseRedirectAfterMeet } from "@/lib/host-entry";
import {
  createMeetingWithBrand,
  parseExternalInviteUrl,
} from "@/lib/meetings";
import {
  clampEmptyTimeoutSec,
  resolveEmptyTimeoutSec,
} from "@/lib/meeting-timeouts";
import { authorizePublicApi } from "@/lib/rooms";
import {
  apiFeatureFields,
  apiFeaturesResponse,
  apiFeaturesToInput,
} from "@/lib/api-features";
import { parseWebhookUrl } from "@/lib/webhook-url";

const schema = z.object({
  /** Título da reunião (obrigatório — definido pela plataforma terceirizada). */
  title: z.string().min(1).max(200),
  access_policy: z.enum(["public", "members", "invite"]).optional(),
  empty_timeout_sec: z.number().int().optional(),
  board_id: z.string().optional(),
  redirect_after_meet: z.string().max(2000).nullable().optional(),
  external_invite_url: z.string().max(2000).nullable().optional(),
  webhook_url: z.string().max(2000).nullable().optional(),
  wait_for_host: z.boolean().optional(),
  mute_mic_on_join: z.boolean().optional(),
  ...apiFeatureFields,
  /** May hosts live stream this meeting? Omitted = LIVE_STREAM_MEETING_DEFAULT. */
  live_stream_enabled: z.boolean().optional(),
});

/**
 * POST /api/v1/rooms/{room_id}/meetings
 * Create a meeting from a predefined room template. Brand is snapshotted from
 * the room; meeting title is provided by the caller. Bearer required.
 */
export async function POST(
  req: NextRequest,
  ctx: { params: Promise<{ room_id: string }> },
) {
  const auth = await authorizePublicApi(req);
  if (!auth.ok) {
    return NextResponse.json({ error: "unauthorized" }, { status: 401 });
  }

  const { room_id: roomId } = await ctx.params;
  if (
    !/^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(
      roomId,
    )
  ) {
    return NextResponse.json({ error: "invalid_room_id" }, { status: 400 });
  }

  let body: z.infer<typeof schema>;
  try {
    body = schema.parse(await req.json());
  } catch (err) {
    return NextResponse.json(
      { error: "invalid_body", detail: String(err) },
      { status: 400 },
    );
  }

  let emptyTimeoutSec: number | null = null;
  try {
    emptyTimeoutSec = clampEmptyTimeoutSec(body.empty_timeout_sec);
  } catch {
    return NextResponse.json(
      {
        error: "empty_timeout_sec_out_of_range",
        detail: "empty_timeout_sec must be between 60 and 86400",
      },
      { status: 400 },
    );
  }

  let redirectAfterMeet: string | null = null;
  try {
    redirectAfterMeet = parseRedirectAfterMeet(body.redirect_after_meet);
  } catch (err) {
    const code = err instanceof Error ? err.message : "redirect_after_meet_invalid";
    return NextResponse.json({ error: code }, { status: 400 });
  }

  const room = await db.query.rooms.findFirst({
    where: eq(rooms.id, roomId),
  });
  if (!room) {
    return NextResponse.json({ error: "room_not_found" }, { status: 404 });
  }

  let webhookUrl: string | null = null;
  try {
    webhookUrl =
      body.webhook_url !== undefined
        ? parseWebhookUrl(body.webhook_url)
        : (room.webhookUrl ?? null);
  } catch (err) {
    const code = err instanceof Error ? err.message : "webhook_url_invalid";
    return NextResponse.json({ error: code }, { status: 400 });
  }

  let externalInviteUrl: string | null = null;
  try {
    externalInviteUrl = parseExternalInviteUrl(body.external_invite_url);
  } catch (err) {
    const code =
      err instanceof Error ? err.message : "external_invite_url_invalid";
    return NextResponse.json({ error: code }, { status: 400 });
  }

  try {
    const accessPolicy =
      body.access_policy ||
      (room.accessPolicy as "public" | "members" | "invite");
    const waitForHost =
      body.wait_for_host !== undefined
        ? body.wait_for_host
        : accessPolicy === "invite";
    const muteMicOnJoin =
      body.mute_mic_on_join !== undefined
        ? body.mute_mic_on_join
        : undefined;

    const { meeting, url, joinPath, hostUrl, hostPath } =
      await createMeetingWithBrand({
        title: body.title.trim(),
        ownerIdentityId: room.ownerIdentityId,
        roomId: room.id,
        boardId: body.board_id ?? room.boardId,
        accessPolicy,
        useIdentityBrand: false,
        emptyTimeoutSec,
        redirectAfterMeet,
        externalInviteUrl,
        webhookUrl,
        waitForHost,
        muteMicOnJoin,
        ...apiFeaturesToInput(body),
        liveStreamEnabled: body.live_stream_enabled,
        issueHostEntry: true,
      });

    return NextResponse.json(
      {
        meeting_id: meeting.id,
        slug: meeting.slug,
        url,
        join_path: joinPath,
        host_url: hostUrl,
        host_path: hostPath,
        access_policy: meeting.accessPolicy,
        title: meeting.title,
        brand_room_id: meeting.roomId,
        empty_timeout_sec: resolveEmptyTimeoutSec(meeting.emptyTimeoutSec),
        redirect_after_meet: meeting.redirectAfterMeet ?? null,
        external_invite_url: meeting.externalInviteUrl ?? null,
        webhook_url: meeting.webhookUrl ?? null,
        wait_for_host: meeting.waitForHost,
        mute_mic_on_join: meeting.muteMicOnJoin !== false,
        ...apiFeaturesResponse(meeting),
        live_stream_enabled: meeting.liveStreamEnabled ?? null,
      },
      { status: 201 },
    );
  } catch (err) {
    const message = err instanceof Error ? err.message : String(err);
    if (message === "room_template_not_found") {
      return NextResponse.json({ error: "room_not_found" }, { status: 404 });
    }
    return NextResponse.json(
      { error: "create_failed", detail: message },
      { status: 500 },
    );
  }
}